llm/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
requests>=2.31.0
pydantic>=2.0.0
langchain>=0.3.0
langchain-core>=0.3.0
grpcio>=1.60.0
grpcio-tools>=1.60.0
